+++ docs/changelog.rst
@@ -32,6 +32,7 @@ fixes
 - default permissions can get duplicated after migration
 - fixed changeset status labels, they now select radio buttons
 - #682 translation difficult for multi-line text
+- #683 fixed difference between messages about not mapped repositories
 
 1.5.0 (**2012-12-12**)
 ----------------------